#88 refactorings - split extensions - remove companion obj in application
