+++ app/src/main/java/dev/leonlatsch/photok/model/repositories/PhotoRepository.kt
@@ -26,8 +26,8 @@ import dev.leonlatsch.photok.model.database.dao.PhotoDao
 import dev.leonlatsch.photok.model.database.entity.Photo
 import dev.leonlatsch.photok.model.database.entity.PhotoType
 import dev.leonlatsch.photok.model.io.EncryptedStorageManager
+import dev.leonlatsch.photok.other.extensions.lazyClose
 import dev.leonlatsch.photok.other.getFileName
-import dev.leonlatsch.photok.other.lazyClose
 import timber.log.Timber
 import java.io.IOException
 import java.io.InputStream
