+++ app/src/main/java/dev/leonlatsch/photok/other/extensions/DialogExtensions.kt
@@ -1,0 +1,27 @@
+/*
+ *   Copyright 2020-2021 Leon Latsch
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package dev.leonlatsch.photok.other.extensions
+
+import androidx.fragment.app.DialogFragment
+import androidx.fragment.app.FragmentManager
+
+/**
+ * Show a DialogFragment with its simple class name as tag.
+ */
+fun DialogFragment.show(fragmentManager: FragmentManager) {
+    this.show(fragmentManager, this::class.simpleName)
+}
