+++ app/src/main/java/dev/leonlatsch/photok/ui/process/base/BaseProcessBottomSheetDialogFragment.kt
@@ -23,9 +23,9 @@ import androidx.core.content.ContextCompat
 import dev.leonlatsch.photok.BR
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.DialogBottomSheetProcessBinding
-import dev.leonlatsch.photok.other.hide
-import dev.leonlatsch.photok.other.show
-import dev.leonlatsch.photok.other.vanish
+import dev.leonlatsch.photok.other.extensions.hide
+import dev.leonlatsch.photok.other.extensions.show
+import dev.leonlatsch.photok.other.extensions.vanish
 import dev.leonlatsch.photok.ui.components.bindings.BindableBottomSheetDialogFragment
 
 /**
