+++ app/src/main/java/dev/leonlatsch/photok/ui/settings/AboutFragment.kt
@@ -25,8 +25,8 @@ import androidx.navigation.fragment.findNavController
 import dev.leonlatsch.photok.BuildConfig
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.FragmentAboutBinding
+import dev.leonlatsch.photok.other.extensions.show
 import dev.leonlatsch.photok.other.openUrl
-import dev.leonlatsch.photok.other.show
 import dev.leonlatsch.photok.ui.components.bindings.BindableFragment
 import dev.leonlatsch.photok.ui.news.NewsDialog
