+++ app/src/main/java/dev/leonlatsch/photok/ui/settings/changepassword/ChangePasswordDialog.kt
@@ -24,9 +24,9 @@ import dagger.hilt.android.AndroidEntryPoint
 import dev.leonlatsch.photok.BR
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.DialogChangePasswordBinding
-import dev.leonlatsch.photok.other.empty
-import dev.leonlatsch.photok.other.hide
-import dev.leonlatsch.photok.other.show
+import dev.leonlatsch.photok.other.extensions.empty
+import dev.leonlatsch.photok.other.extensions.hide
+import dev.leonlatsch.photok.other.extensions.show
 import dev.leonlatsch.photok.security.PasswordUtils
 import dev.leonlatsch.photok.ui.components.Dialogs
 import dev.leonlatsch.photok.ui.components.bindings.BindableDialogFragment
