+++ app/src/main/java/dev/leonlatsch/photok/ui/settings/hideapp/ToggleAppVisibilityViewModel.kt
@@ -24,7 +24,7 @@ import androidx.databinding.Bindable
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.leonlatsch.photok.BR
 import dev.leonlatsch.photok.R
-import dev.leonlatsch.photok.other.empty
+import dev.leonlatsch.photok.other.extensions.empty
 import dev.leonlatsch.photok.settings.Config
 import dev.leonlatsch.photok.ui.components.bindings.ObservableViewModel
 import javax.inject.Inject
